Data-Pipeline/requirements.txt
pytest==8.3.2
requests==2.32.3
tqdm==4.66.5
pillow==10.4.0
pyyaml==6.0.2
dvc